+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Updated Bridgeless to use minimum swap amount provided by API
+
 ## 2.36.0 (2025-10-16)
 
 - added: (Thorchain) Support for Tron